qcom: subsys-pil-tz: Add property for setting clock frequencies
Add a new property, "qcom,'clock_name'-freq" for choosing the
frequency to set for the active or proxy clocks. If the
frequency property for a clock isnt added, then the clock
would be set to XO frequency by default.

This change for adding support to configure these
frequencies from dt is required for scm clocks.

Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>